update: add autho to userrel ctrl and route

+++ controllers/usersRelationship.controller.js
@@ -7,6 +7,14 @@ const UsersRelationship = require('../models/UsersRelationship')
 async function getAllUserRel(req,res){
     try {
     // TODO: add authorization here--⚠️⚠️
+    if(req.user.role !== "admin" ) {
+        const userRel = await UsersRelationship.findAll({
+            where:{
+                userId: req.user.id
+            }
+        })
+       return res.json(userRel) 
+    }
 
         const userRel = await UsersRelationship.findAll()
         res.json(userRel)
@@ -20,8 +28,10 @@ async function getAllUserRel(req,res){
 async function getUserRelById(req,res){
     try {
     // TODO: add authorization here--⚠️⚠️
-        
     const userRel = await UsersRelationship.findByPk(parseInt(req.params.userRelId))
+        
+    if(userRel.userId !== req.user.id && req.user.role !== "admin") throw `Restricted access: Cannot view other user relationship`
+
     res.json(userRel)
         
     } catch (error) {
@@ -49,6 +59,9 @@ async function createUserRel(req,res){
 async function updateUserRel(req,res){
     try {
     // TODO: add authorization here--⚠️⚠️
+    const userRel = await UsersRelationship.findByPk(parseInt(req.params.userRelId))
+        
+    if(userRel.userId !== req.user.id && req.user.role !== "admin") throw `Restricted access: Cannot update other user relationship`
       
     const updatedUserRel = await UsersRelationship.update(req.body, {
         where: {
@@ -66,6 +79,9 @@ async function updateUserRel(req,res){
 async function deleteUserRel(req,res){
     try {
     // TODO: add authorization here--⚠️⚠️
+    const userRel = await UsersRelationship.findByPk(parseInt(req.params.userRelId))
+        
+    if(userRel.userId !== req.user.id && req.user.role !== "admin") throw `Restricted access: Cannot delete other user relationship`
        
     const deletedUserRel = await UsersRelationship.destroy({
         where:{
